encapsulate logic with force float32 encoding in packer_float32_decorator class
