+++ runtime/msgpack-serialization.cpp
@@ -6,8 +6,6 @@
 
 #include "msgpack/adaptors.h"
 
-uint32_t serialize_as_float32{0};
-
 namespace msgpack::adaptor {
 
 size_t CheckInstanceDepth::depth = 0;
