+++ runtime/msgpack/packer.h
@@ -11,6 +11,7 @@
 
 #include <cstddef>
 #include <cstdint>
+#include <tuple>
 
 #include "common/mixin/not_copyable.h"
 
@@ -67,6 +68,58 @@ private:
   void append_buffer(const char *buf, size_t len) noexcept;
 
   Stream &stream_;
+};
+
+class packer_float32_decorator {
+public:
+  template<class StreamT, class T>
+  static void pack_value(bool as_float32, packer<StreamT> &packer, const T& value) {
+    if (as_float32) {
+      ++serialize_as_float32_;
+      pack_value(packer, value);
+      --serialize_as_float32_;
+    } else {
+      pack_value(packer, value);
+    }
+  }
+
+  template<class StreamT, class T>
+  static void pack_value(msgpack::packer<StreamT> &packer, const T& value) {
+    packer.pack(value);
+  }
+
+  template<class StreamT>
+  static void pack_value(msgpack::packer<StreamT> &packer, double value) {
+    if (serialize_as_float32_ > 0) {
+      packer.pack(static_cast<float>(value));
+    } else {
+      packer.pack(value);
+    }
+  }
+
+private:
+  template<size_t N>
+  struct PackValueHelper {
+    template<class StreamT, class TupleT>
+    static void convert(packer<StreamT> &packer, const TupleT &value) {
+      PackValueHelper<N - 1>::convert(packer, value);
+      pack_value(packer, std::get<N - 1>(value));
+    }
+  };
+
+  template<>
+  struct PackValueHelper<0> {
+    template<class StreamT, class TupleT>
+    static void convert(packer<StreamT> &, const TupleT &) {}
+  };
+
+  template<class StreamT, class... Args>
+  static void pack_value(packer<StreamT> &packer, const std::tuple<Args...> &value) {
+    packer.pack_array(sizeof...(Args));
+    PackValueHelper<sizeof...(Args)>::convert(packer, value);
+  }
+
+  static uint32_t serialize_as_float32_;
 };
 
 } // namespace msgpack
